Test fixes, move more tests to op validation suite; get all ignored or passing
