+++ nd4j/nd4j-backends/nd4j-tests/src/test/java/org/nd4j/autodiff/opvalidation/LayerOpValidation.java
@@ -1,5 +1,6 @@
 package org.nd4j.autodiff.opvalidation;
 
+import lombok.val;
 import org.junit.Test;
 import org.nd4j.autodiff.OpValidationSuite;
 import org.nd4j.autodiff.samediff.SDVariable;
@@ -19,9 +20,8 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.fail;
+import static org.junit.Assert.*;
+import static org.junit.Assert.assertArrayEquals;
 
 public class LayerOpValidation extends BaseOpValidation {
     public LayerOpValidation(Nd4jBackend backend) {
@@ -578,5 +578,383 @@ public class LayerOpValidation extends BaseOpValidation {
         assertEquals(failed.toString(), 0, failed.size());
     }
 
+
+    @Test
+    public void testDepthWiseConv2dBasic() {
+        OpValidationSuite.ignoreFailing();
+
+        int nIn = 3;
+        int depthWise = 4;
+        int kH = 2;
+        int kW = 2;
+
+        int mb = 3;
+        int imgH = 28;
+        int imgW = 28;
+
+
+        SameDiff sd = SameDiff.create();
+        INDArray depthWeightArr = Nd4j.create(depthWise, nIn, kH, kW);
+
+        INDArray bArr = Nd4j.create(1, depthWise * nIn);
+        INDArray inArr = Nd4j.create(mb, nIn, imgH, imgW);
+
+        SDVariable in = sd.var("in", inArr);
+        SDVariable dW = sd.var("dW", depthWeightArr);
+        SDVariable b = sd.var("b", bArr);
+
+        SDVariable[] vars = new SDVariable[]{in, dW, b};
+
+        Conv2DConfig c = Conv2DConfig.builder()
+                .kh(kH).kw(kW)
+                .ph(0).pw(0)
+                .sy(1).sx(1)
+                .dh(1).dw(1)
+                .isSameMode(false)
+                .build();
+
+        SDVariable out = sd.sconv2d(vars, c);
+        out = sd.tanh("out", out);
+
+        INDArray outArr = sd.execAndEndResult();
+        //Expected output size: out = (in - k + 2*p)/s + 1 = (28-2+0)/1+1 = 27
+        val outShape = outArr.shape();
+        assertArrayEquals(new long[]{mb, depthWise * nIn, 27, 27}, outShape);
+    }
+
+    @Test
+    public void testSeparableConv2dBasic() {
+        OpValidationSuite.ignoreFailing();
+        int nIn = 3;
+        int nOut = 4;
+        int kH = 2;
+        int kW = 2;
+
+        int mb = 3;
+        int imgH = 28;
+        int imgW = 28;
+
+        int depthWise = 3;
+
+        SameDiff sd = SameDiff.create();
+        INDArray depthWeightArr = Nd4j.create(depthWise, nIn, kH, kW);
+        INDArray pointWeightArr = Nd4j.create(nOut, depthWise, 1, 1);
+
+        INDArray bArr = Nd4j.create(1, nOut);
+        INDArray inArr = Nd4j.create(mb, nIn, imgH, imgW);
+
+        SDVariable in = sd.var("in", inArr);
+        SDVariable dW = sd.var("dW", depthWeightArr);
+        SDVariable pW = sd.var("pW", pointWeightArr);
+        SDVariable b = sd.var("b", bArr);
+
+        SDVariable[] vars = new SDVariable[]{in, dW, pW, b};
+
+        Conv2DConfig c = Conv2DConfig.builder()
+                .kh(kH).kw(kW)
+                .ph(0).pw(0)
+                .sy(1).sx(1)
+                .dh(1).dw(1)
+                .isSameMode(false)
+                .build();
+
+        SDVariable out = sd.sconv2d(vars, c);
+        out = sd.tanh("out", out);
+
+        INDArray outArr = sd.execAndEndResult();
+        //Expected output size: out = (in - k + 2*p)/s + 1 = (28-2+0)/1+1 = 27
+        val outShape = outArr.shape();
+        assertArrayEquals(new long[]{mb, nOut, 27, 27}, outShape);
+    }
+
+    @Test
+    public void testDeconv2dBasic() {
+        OpValidationSuite.ignoreFailing();
+
+        int nIn = 3;
+        int nOut = 4;
+        int kH = 2;
+        int kW = 2;
+
+        int mb = 3;
+        int imgH = 28;
+        int imgW = 28;
+
+        SameDiff sd = SameDiff.create();
+        INDArray wArr = Nd4j.create(nOut, nIn, kH, kW);
+        INDArray bArr = Nd4j.create(1, nOut);
+        INDArray inArr = Nd4j.create(mb, nIn, imgH, imgW);
+
+        SDVariable in = sd.var("in", inArr);
+        SDVariable w = sd.var("W", wArr);
+        SDVariable b = sd.var("b", bArr);
+
+        SDVariable[] vars = new SDVariable[]{in, w, b};
+
+        DeConv2DConfig deconv = DeConv2DConfig.builder()
+                .kX(kH).kY(kW)
+                .pX(0).pY(0)
+                .sX(1).sY(1)
+                .dX(1).dY(1)
+                .isSameMode(false)
+                .build();
+
+        SDVariable out = sd.deconv2d(vars, deconv);
+        out = sd.tanh("out", out);
+
+        INDArray outArr = sd.execAndEndResult();
+        //Expected output size: out = (in + k + 2*p)/ s - 1 = (28 + 2+0)/1 - 1 = 29
+        val outShape = outArr.shape();
+        assertArrayEquals(new long[]{mb, nOut, 29, 29}, outShape);
+    }
+
+
+    @Test
+    public void testConv2dBasic() {
+        int nIn = 3;
+        int nOut = 4;
+        int kH = 2;
+        int kW = 2;
+
+        int mb = 3;
+        int imgH = 28;
+        int imgW = 28;
+
+        SameDiff sd = SameDiff.create();
+        INDArray wArr = Nd4j.create(nOut, nIn, kH, kW); //As per DL4J
+        INDArray bArr = Nd4j.create(1, nOut);
+        INDArray inArr = Nd4j.create(mb, nIn, imgH, imgW);
+
+        SDVariable in = sd.var("in", inArr);
+        SDVariable w = sd.var("W", wArr);
+        SDVariable b = sd.var("b", bArr);
+
+        //Order: https://github.com/deeplearning4j/libnd4j/blob/6c41ea5528bb1f454e92a9da971de87b93ff521f/include/ops/declarable/generic/convo/conv2d.cpp#L20-L22
+        //in, w, b - bias is optional
+        SDVariable[] vars = new SDVariable[]{in, w, b};
+
+        Conv2DConfig c = Conv2DConfig.builder()
+                .kh(kH).kw(kW)
+                .ph(0).pw(0)
+                .sy(1).sx(1)
+                .dh(1).dw(1)
+                .isSameMode(false)
+                .build();
+
+        SDVariable out = sd.conv2d(vars, c);
+        out = sd.tanh("out", out);
+
+        INDArray outArr = sd.execAndEndResult();
+        //Expected output size: out = (in - k + 2*p)/s + 1 = (28-2+0)/1+1 = 27
+        val outShape = outArr.shape();
+        assertArrayEquals(new long[]{mb, nOut, 27, 27}, outShape);
+        // sd.execBackwards(); // TODO: test failing here
+    }
+
+    @Test
+    public void testMaxPooling2dBasic() {
+        int nIn = 3;
+        int kH = 2;
+        int kW = 2;
+
+        int mb = 3;
+        int imgH = 28;
+        int imgW = 28;
+
+        SameDiff sd = SameDiff.create();
+        INDArray inArr = Nd4j.create(mb, nIn, imgH, imgW);
+
+        SDVariable in = sd.var("in", inArr);
+
+        Pooling2DConfig pooling2DConfig = Pooling2DConfig.builder()
+                .kh(kH).kw(kW)
+                .ph(0).pw(0)
+                .sy(1).sx(1)
+                .dh(1).dw(1)
+                .isSameMode(false)
+                .build();
+
+        SDVariable out = sd.maxPooling2d(in, pooling2DConfig);
+        out = sd.tanh("out", out);
+
+        INDArray outArr = sd.execAndEndResult();
+        val outShape = outArr.shape();
+        // oH = (iH - (kH + (kH-1)*(dH-1)) + 2*pH)/sH + 1;
+        assertArrayEquals(new long[]{mb, nIn, 27, 27}, outShape);
+    }
+
+    @Test
+    public void testAvgPooling2dBasic() {
+        int nIn = 3;
+        int kH = 2;
+        int kW = 2;
+
+        int mb = 3;
+        int imgH = 28;
+        int imgW = 28;
+
+        SameDiff sd = SameDiff.create();
+        INDArray inArr = Nd4j.create(mb, nIn, imgH, imgW);
+
+        SDVariable in = sd.var("in", inArr);
+
+        Pooling2DConfig pooling2DConfig = Pooling2DConfig.builder()
+                .kh(kH).kw(kW)
+                .ph(0).pw(0)
+                .sy(1).sx(1)
+                .dh(1).dw(1)
+                .isSameMode(false)
+                .build();
+
+        SDVariable out = sd.avgPooling2d(in, pooling2DConfig);
+        out = sd.tanh("out", out);
+
+        INDArray outArr = sd.execAndEndResult();
+        val outShape = outArr.shape();
+        // oH = (iH - (kH + (kH-1)*(dH-1)) + 2*pH)/sH + 1;
+        assertArrayEquals(new long[]{mb, nIn, 27, 27}, outShape);
+    }
+
+    @Test
+    public void testAvgPooling3dBasic() {
+        OpValidationSuite.ignoreFailing();
+        int nIn = 3;
+        int kH = 2;
+        int kW = 2;
+        int kD = 2;
+
+        int mb = 3;
+        int imgH = 28;
+        int imgW = 28;
+        int imgD = 28;
+
+        SameDiff sd = SameDiff.create();
+        INDArray inArr = Nd4j.create(mb, nIn, imgD, imgH, imgW);
+
+        SDVariable in = sd.var("in", inArr);
+
+        Pooling3DConfig pooling3DConfig = Pooling3DConfig.builder()
+                .kH(kH).kW(kW).kT(kD)
+                .pH(0).pH(0).pT(0)
+                .sH(1).sW(1).sT(1)
+                .dilationH(0).dilationW(0).dilationT(0)
+                .ceilingMode(false)
+                .build();
+
+        SDVariable out = sd.avgPooling3d(in, pooling3DConfig);
+        out = sd.tanh("out", out);
+
+        INDArray outArr = sd.execAndEndResult();
+        val outShape = outArr.shape();
+        // oH = (iH - (kH + (kH-1)*(dH-1)) + 2*pH)/sH + 1;
+        assertArrayEquals(new long[]{mb, nIn, 27, 27, 27}, outShape);
+    }
+
+    @Test
+    public void testMaxPooling3dBasic() {
+        OpValidationSuite.ignoreFailing();
+        int nIn = 3;
+        int kH = 2;
+        int kW = 2;
+        int kD = 2;
+
+        int mb = 3;
+        int imgH = 28;
+        int imgW = 28;
+        int imgD = 28;
+
+        SameDiff sd = SameDiff.create();
+        INDArray inArr = Nd4j.create(mb, nIn, imgD, imgH, imgW);
+
+        SDVariable in = sd.var("in", inArr);
+
+        Pooling3DConfig pooling3DConfig = Pooling3DConfig.builder()
+                .kH(kH).kW(kW).kT(kD)
+                .pH(0).pH(0).pT(0)
+                .sH(1).sW(1).sT(1)
+                .dilationH(0).dilationW(0).dilationT(0)
+                .ceilingMode(false)
+                .build();
+
+        SDVariable out = sd.maxPooling3d(in, pooling3DConfig);
+        out = sd.tanh("out", out);
+
+        INDArray outArr = sd.execAndEndResult();
+        val outShape = outArr.shape();
+        // oH = (iH - (kH + (kH-1)*(dH-1)) + 2*pH)/sH + 1;
+        assertArrayEquals(new long[]{mb, nIn, 27, 27, 27}, outShape);
+    }
+
+    @Test
+    public void testConv1dBasic() {
+        int nIn = 3;
+        int nOut = 4;
+        int k = 2;
+        int mb = 3;
+        int img = 28;
+
+        SameDiff sd = SameDiff.create();
+        INDArray wArr = Nd4j.create(nOut, nIn, k);
+        INDArray inArr = Nd4j.create(mb, nIn, img);
+
+        SDVariable in = sd.var("in", inArr);
+        SDVariable w = sd.var("W", wArr);
+
+        SDVariable[] vars = new SDVariable[]{in, w};
+
+        Conv1DConfig conv1DConfig = Conv1DConfig.builder()
+                .k(k).p(0).s(1)
+                .isSameMode(false)
+                .build();
+
+        SDVariable out = sd.conv1d(vars, conv1DConfig);
+        out = sd.tanh("out", out);
+
+        INDArray outArr = sd.execAndEndResult();
+        INDArray iOut = out.getArr();
+        //Expected output size: out = (in - k + 2*p)/s + 1 = (28-2+0)/1+1 = 27
+        val outShape = outArr.shape();
+        assertArrayEquals(new long[]{mb, nOut, 27}, outShape);
+    }
+
+
+    @Test
+    public void testConv3dBasic() {
+        OpValidationSuite.ignoreFailing();
+        int nIn = 3;
+        int nOut = 4;
+        int kH = 2;
+        int kW = 2;
+        int kT = 2;
+
+        int mb = 3;
+        int imgH = 28;
+        int imgW = 28;
+        int imgT = 28;
+
+        SameDiff sd = SameDiff.create();
+        INDArray wArr = Nd4j.create(nOut, nIn, kT, kH, kW); //As per DL4J
+        INDArray bArr = Nd4j.create(1, nOut);
+        INDArray inArr = Nd4j.create(mb, nIn, imgT, imgH, imgW);
+
+        SDVariable in = sd.var("in", inArr);
+        SDVariable w = sd.var("W", wArr);
+        SDVariable b = sd.var("b", bArr);
+
+        Conv3DConfig conv3DConfig = Conv3DConfig.builder()
+                .kH(kH).kW(kW).kT(kT)
+                .dilationH(1).dilationW(1).dilationT(1)
+                .isValidMode(false)
+                .biasUsed(false)
+                .build();
+
+        SDVariable out = sd.conv3d(in, w, b, conv3DConfig);
+        out = sd.tanh("out", out);
+
+        INDArray outArr = sd.execAndEndResult();
+        //Expected output size: out = (in - k)/d + 1 = (28-2+0)/1+1 = 27
+        val outShape = outArr.shape();
+        assertArrayEquals(new long[]{mb, nOut, 27, 27, 27}, outShape);
+    }
 
 }
