+++ nd4j/nd4j-backends/nd4j-tests/src/test/java/org/nd4j/autodiff/opvalidation/MiscOpValidation.java
@@ -1013,6 +1013,7 @@ public class MiscOpValidation extends BaseOpValidation {
 
     @Test
     public void testZerosOnesLike(){
+        OpValidationSuite.ignoreFailing();
         Nd4j.getRandom().setSeed(12345);
 
         List<int[]> shapes = Arrays.asList(new int[0], new int[]{3}, new int[]{3,4}, new int[]{3,4,5});
@@ -1164,5 +1165,23 @@ public class MiscOpValidation extends BaseOpValidation {
         }
 
         assertEquals(failed.toString(), 0, failed.size());
+    }
+
+    @Test
+    public void testInplaceSubi() {
+        OpValidationSuite.ignoreFailing();
+        SameDiff sameDiffOuter = SameDiff.create();
+        Map<String, INDArray> params = new HashMap<>();
+        params.put("x", Nd4j.ones(4));
+        sameDiffOuter.defineFunction("inplacesubi", new SameDiff.SameDiffFunctionDefinition() {
+            @Override
+            public SDVariable[] define(SameDiff sameDiff, Map<String, INDArray> inputs, SDVariable[] variableInputs) {
+                SDVariable inplace = sameDiff.var("x", inputs.get("x"));
+                return new SDVariable[]{inplace.subi(1.0)};
+            }
+        }, params);
+
+        sameDiffOuter.getFunction("inplacesubi").eval(params);
+        assertEquals(Nd4j.zeros(4), params.get("x"));
     }
 }
