+++ nd4j/nd4j-backends/nd4j-tests/src/test/java/org/nd4j/autodiff/opvalidation/ShapeOpValidation.java
@@ -20,6 +20,7 @@ import org.nd4j.linalg.api.ops.impl.shape.Unstack;
 import org.nd4j.linalg.checkutil.NDArrayCreationUtil;
 import org.nd4j.linalg.factory.Nd4j;
 import org.nd4j.linalg.factory.Nd4jBackend;
+import org.nd4j.linalg.indexing.NDArrayIndex;
 import org.nd4j.linalg.ops.transforms.Transforms;
 import org.nd4j.linalg.primitives.Pair;
 import org.nd4j.linalg.primitives.Triple;
@@ -29,6 +30,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
+import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 import static org.nd4j.linalg.indexing.NDArrayIndex.*;
@@ -1054,6 +1056,7 @@ public class ShapeOpValidation extends BaseOpValidation {
 
     @Test
     public void testMeshGrid(){
+        OpValidationSuite.ignoreFailing();
 
         List<String> failed = new ArrayList<>();
 
@@ -1219,5 +1222,144 @@ public class ShapeOpValidation extends BaseOpValidation {
         INDArray arr2 = Nd4j.concat(0, arr, arr);  // (1, 4), (1, 4) -> (2, 4)
         INDArray expected = Nd4j.concat(1, arr2, arr2);  // (2, 4), (2, 4) -> (2, 8)
         assertEquals(expected, result.eval());
+    }
+
+    @Test
+    public void testBroadcast() {
+        OpValidationSuite.ignoreFailing();
+        SameDiff sd = SameDiff.create();
+        SDVariable in = sd.var("in", Nd4j.rand(3, 4));
+        SDVariable broadcast = sd.f().broadcast(in, 3, 4, 5);
+
+        INDArray out = sd.execAndEndResult();
+        assertArrayEquals(new long[]{3, 4, 5}, out.shape());
+
+        for (int i = 0; i < 5; i++) {
+            assertEquals(in.getArr(), out.get(all(), all(), point(i)));
+        }
+    }
+
+
+    @Test
+    public void testSlice2d() {
+        INDArray inArr = Nd4j.linspace(1, 12, 12).reshape('c', 3, 4);
+
+        SameDiff sd = SameDiff.create();
+        SDVariable in = sd.var("in", inArr);
+        SDVariable slice_full = sd.slice(in, new int[]{0, 0}, new int[]{3, 4});
+        SDVariable subPart = sd.slice(in, new int[]{1, 2}, new int[]{2, 2});
+
+        sd.execAndEndResult();
+
+        assertEquals(inArr, slice_full.getArr());
+        assertEquals(inArr.get(interval(1, 3), interval(2, 4)), subPart.getArr());
+    }
+
+
+    @Test
+    public void testSlice3d() {
+        INDArray inArr = Nd4j.linspace(1, 60, 60).reshape('c', 3, 4, 5);
+
+        SameDiff sd = SameDiff.create();
+        SDVariable in = sd.var("in", inArr);
+        SDVariable slice_full = sd.slice(in, new int[]{0, 0, 0}, new int[]{3, 4, 5});
+        SDVariable subPart = sd.slice(in, new int[]{1, 2, 3}, new int[]{2, 2, 1});
+
+        sd.execAndEndResult();
+
+        assertEquals(inArr, slice_full.getArr());
+        assertEquals(inArr.get(interval(1, 3), interval(2, 4), interval(3, 4)), subPart.getArr());
+    }
+
+    @Test
+    public void testStridedSlice2dBasic() {
+        INDArray inArr = Nd4j.linspace(1, 12, 12).reshape('c', 3, 4);
+
+        SameDiff sd = SameDiff.create();
+        SDVariable in = sd.var("in", inArr);
+        SDVariable slice_full = sd.stridedSlice(in, new int[]{0, 0}, new int[]{3, 4}, new int[]{1, 1});
+        SDVariable subPart = sd.stridedSlice(in, new int[]{1, 2}, new int[]{3, 4}, new int[]{1, 1});
+        SDVariable subPart2 = sd.stridedSlice(in, new int[]{0, 0}, new int[]{4, 5}, new int[]{2, 2});
+
+        sd.execAndEndResult();
+
+        assertEquals(inArr, slice_full.getArr());
+        assertEquals(inArr.get(interval(1, 3), interval(2, 4)), subPart.getArr());
+        assertEquals(inArr.get(interval(0, 2, 4), interval(0, 2, 5)), subPart2.getArr());
+    }
+
+
+    @Test
+    public void testStridedSliceBeginEndMask() {
+        INDArray inArr = Nd4j.linspace(1, 12, 12).reshape('c', 3, 4);
+
+        SameDiff sd = SameDiff.create();
+        SDVariable in = sd.var("in", inArr);
+        SDVariable slice1 = sd.stridedSlice(in, new int[]{-999, 0}, new int[]{2, 4}, new int[]{1, 1}, 1 << 1, 0, 0, 0, 0);
+        SDVariable slice2 = sd.stridedSlice(in, new int[]{1, 0}, new int[]{-999, 4}, new int[]{1, 1}, 0, 1, 0, 0, 0);
+
+        sd.execAndEndResult();
+
+        assertEquals(inArr.get(NDArrayIndex.interval(0, 2), NDArrayIndex.all()), slice1.getArr());
+        assertEquals(inArr.get(NDArrayIndex.interval(1, 3), NDArrayIndex.all()), slice2.getArr());
+    }
+
+    @Test
+    public void testStridedSliceEllipsisMask() {
+        INDArray inArr = Nd4j.linspace(1, 60, 60).reshape('c', 3, 4, 5);
+        SameDiff sd = SameDiff.create();
+        SDVariable in = sd.var("in", inArr);
+
+        //[1:3,...] -> [1:3,:,:]
+        SDVariable slice = sd.stridedSlice(in, new int[]{1}, new int[]{3}, new int[]{1}, 0, 0, 1 << 1, 0, 0);
+        //[1:3,...,1:4] -> [1:3,:,1:4]
+        SDVariable slice2 = sd.stridedSlice(in, new int[]{1, 1}, new int[]{3, 4}, new int[]{1, 1}, 0, 0, 1 << 1, 0, 0);
+
+        sd.execAndEndResult();
+
+        assertEquals(inArr.get(interval(1, 3), all(), all()), slice.getArr());
+        assertEquals(inArr.get(interval(1, 3), all(), all()), slice2.getArr());
+    }
+
+    @Test
+    public void testStridedSliceNewAxisMask() {
+        OpValidationSuite.ignoreFailing();
+        INDArray inArr = Nd4j.linspace(1, 60, 60).reshape('c', 3, 4, 5);
+        SameDiff sd = SameDiff.create();
+        SDVariable in = sd.var("in", inArr);
+        SDVariable slice = sd.stridedSlice(in, new int[]{-999, 0, 0, 0}, new int[]{-999, 3, 4, 5}, new int[]{-999, 1, 1, 1}, 0, 0, 0, 1, 0);
+
+        INDArray out = sd.execAndEndResult();
+
+        assertArrayEquals(new long[]{1, 3, 4, 5}, inArr.shape());
+        assertEquals(inArr, out.get(point(0), all(), all(), all()));
+    }
+
+    @Test
+    public void testStridedSliceNewAxisMask2() {
+        INDArray inArr = Nd4j.linspace(1, 60, 60).reshape('c', 3, 4, 5);
+        SameDiff sd = SameDiff.create();
+        SDVariable in = sd.var("in", inArr);
+        SDVariable slice = sd.stridedSlice(in, new int[]{1, 1, -999, 1}, new int[]{3, 3, -999, 4}, new int[]{1, 1, -999, 1}, 0, 0, 0, 1 << 2, 0);
+        INDArray out = sd.execAndEndResult();
+
+        assertArrayEquals(new long[]{2, 2, 1, 3}, slice.getArr().shape());
+    }
+
+    @Test
+    public void testStridedSliceShrinkAxisMask() {
+
+        INDArray inArr = Nd4j.linspace(1, 60, 60).reshape('c', 3, 4, 5);
+        SameDiff sd = SameDiff.create();
+        SDVariable in = sd.var("in", inArr);
+        SDVariable slice = sd.stridedSlice(in, new int[]{0, 0, 0}, new int[]{-999, 4, 5}, new int[]{1, 1, 1}, 0, 0, 0, 0, 1);
+        SDVariable slice2 = sd.stridedSlice(in, new int[]{2, 0, 0}, new int[]{-999, 4, 5}, new int[]{1, 1, 1}, 0, 0, 0, 0, 1);
+        SDVariable slice3 = sd.stridedSlice(in, new int[]{1, 2, 1}, new int[]{-999, -999, 5}, new int[]{1, 1, 1}, 0, 0, 0, 0, 1 | 1 << 1);
+
+        sd.execAndEndResult();
+
+        assertEquals(inArr.get(point(0), all(), all()), slice.getArr());
+        assertEquals(inArr.get(point(2), all(), all()), slice2.getArr());
+        assertEquals(inArr.get(point(1), point(2), interval(1, 5)), slice3.getArr());
     }
 }
